Homework

+++ src/main/java/ie/atu/week5/customerapp/CustomerService.java
@@ -1,9 +1,14 @@
 package ie.atu.week5.customerapp;
 
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CustomerService {
@@ -18,18 +23,25 @@ public class CustomerService {
         return customerRepository.findAll();
     }
 
-    public List<Customer> getCustomerById(String id){
-        customerRepository.findById(id);
-        return customerRepository.findAll();
+    public ResponseEntity<Customer> getCustomerById(String id) {
+        Optional<Customer> customer = customerRepository.findById(id);
+        if (customer.isPresent()) {
+            return ResponseEntity.ok(customer.get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
-    public void create(Customer customer){
-        customerRepository.save(customer);
+    public ResponseEntity<Customer> createCustomer(Customer customer) {
+        Customer savedCustomer = customerRepository.save(customer);
+        return ResponseEntity.ok(savedCustomer);
     }
 
-
-    public List<Customer> delete(String id){
-        customerRepository.deleteById(id);
-        return customerRepository.findAll();
+    public ResponseEntity<Void> deleteCustomer(String id) {
+        if (customerRepository.existsById(id)) {
+            customerRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
